@@ -1,12 +1,17 @@
-#include "../interfaces/TapeConfigFactory.h"
+// FileTapeConfigFactory.h
+#pragma once
 
+#include "../interfaces/TapeConfigFactoryInterface.h"
 #include <string>
 
-class FileTapeConfigFactory : public TapeConfigFactory {
+class FileTapeConfigFactory : public TapeConfigFactoryInterface {
 public:
-  explicit FileTapeConfigFactory(const std::string &filename);
-  TapeConfig createConfig() const final;
+  explicit FileTapeConfigFactory(std::string filename);
+  TapeConfig create() const final;
 
 private:
-  std::string configFile;
+  std::string configFile_;
+
+  void parseLine(const std::string &line, TapeConfig &config) const;
+  void validateConfig(const TapeConfig &config) const;
 };
